+++ Tests/SentryTests/Helper/TestFileManagerDelegate.swift
@@ -4,7 +4,7 @@ import SentryTestUtils
 class TestFileManagerDelegate: NSObject, SentryFileManagerDelegate {
     
     var envelopeItemsDeleted = Invocations<SentryDataCategory>()
-    func envelopeItemDeleted(_ dataCategory: SentryDataCategory) {
+    func envelopeItemDeleted(_ envelopeItem: SentryEnvelopeItem, with dataCategory: SentryDataCategory) {
         envelopeItemsDeleted.record(dataCategory)
     }
 }