+++ Sources/Sentry/SentrySpotlightTransport.m
@@ -96,11 +96,19 @@ SentrySpotlightTransport ()
     // Empty on purpose
 }
 
+- (void)recordLostEvent:(SentryDataCategory)category
+                 reason:(SentryDiscardReason)reason
+               quantity:(NSUInteger)quantity
+{
+    // Empty on purpose
+}
+
 #if defined(TEST) || defined(TESTCI) || defined(DEBUG)
 - (void)setStartFlushCallback:(nonnull void (^)(void))callback
 {
     // Empty on purpose
 }
+
 #endif // defined(TEST) || defined(TESTCI) || defined(DEBUG)
 
 @end