+++ Sources/Sentry/SentryTransportAdapter.m
@@ -102,6 +102,15 @@ SentryTransportAdapter ()
     }
 }
 
+- (void)recordLostEvent:(SentryDataCategory)category
+                 reason:(SentryDiscardReason)reason
+               quantity:(NSUInteger)quantity
+{
+    for (id<SentryTransport> transport in self.transports) {
+        [transport recordLostEvent:category reason:reason quantity:quantity];
+    }
+}
+
 - (void)flush:(NSTimeInterval)timeout
 {
     for (id<SentryTransport> transport in self.transports) {