+++ Tests/SentryTests/Networking/RateLimits/TestEnvelopeRateLimitDelegate.swift
@@ -4,7 +4,7 @@ import SentryTestUtils
 class TestEnvelopeRateLimitDelegate: NSObject, SentryEnvelopeRateLimitDelegate {
     
     var envelopeItemsDropped = Invocations<SentryDataCategory>()
-    func envelopeItemDropped(_ dataCategory: SentryDataCategory) {
+    func envelopeItemDropped(_ envelopeItem: SentryEnvelopeItem, with dataCategory: SentryDataCategory) {
         envelopeItemsDropped.record(dataCategory)
     }
 }