+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Features
 
 - Add `reportAccessibilityIdentifier` option (#4183)
+- Record dropped spans (#4172)
+
+### Fixes
+
+- Session replay crash when writing the replay (#4186)
 
 ## 8.31.1